Builds the table of GSEA statistics to present to the user.

Creates a table gene set statistics genereated from a GSEA `method` for all gene sets with an FDR (by collection) less than given threshold.

constructGseaResultTable(
  mg,
  method = sparrow::resultNames(mg)[1L],
  fdr = 0.1,
  prioritize = NULL
)

Arguments

mg

`SparrowResult` object

method

the method to show statistics for

fdr

the FDR cut off to present statistics for

prioritize

the preferred collections to put at the top of the list. The collection column of the table is turned into a factor for more useful display with datatable's filter. Specifcying collections here will put those collections at the front of the factor levels and therofre prioritize their display in the select dropdown for the filter

Value

a data.table of the statistics that match the filtering criteria. A 0-row data.table is returned if nothing passes.

Examples

sres <- sparrow::exampleSparrowResult()
constructGseaResultTable(sres)
#>     collection                                                    name active
#>          <ord>                                                  <char> <lgcl>
#>  1:         c2                      LOPEZ_MESOTELIOMA_SURVIVAL_TIME_UP   TRUE
#>  2:         c2                  SOTIRIOU_BREAST_CANCER_GRADE_1_VS_3_UP   TRUE
#>  3:         c2              PUJANA_BREAST_CANCER_WITH_BRCA1_MUTATED_UP   TRUE
#>  4:         c2                    REACTOME_MRNA_SPLICING_MINOR_PATHWAY   TRUE
#>  5:         c7                                 GSE3982_BCELL_VS_TH2_DN   TRUE
#>  6:         c6                                           E2F3_UP.V1_UP   TRUE
#>  7:         c7             GSE13485_CTRL_VS_DAY3_YF17D_VACCINE_PBMC_DN   TRUE
#>  8:         c2                       CAIRO_PML_TARGETS_BOUND_BY_MYC_UP   TRUE
#>  9:         c2                     MOREAUX_MULTIPLE_MYELOMA_BY_TACI_DN   TRUE
#> 10:         c6                                        TBK1.DN.48HRS_DN   TRUE
#> 11:         c2                          VANTVEER_BREAST_CANCER_ESR1_DN   TRUE
#> 12:         c6                                           CAMP_UP.V1_DN   TRUE
#> 13:         c6                           KRAS.600.LUNG.BREAST_UP.V1_DN   TRUE
#> 14:         c7   GSE22886_IGG_IGA_MEMORY_BCELL_VS_BLOOD_PLASMA_CELL_UP   TRUE
#> 15:         c6                                    KRAS.BREAST_UP.V1_DN   TRUE
#> 16:         c2                  SCHAEFFER_PROSTATE_DEVELOPMENT_48HR_DN   TRUE
#> 17:         c6                               KRAS.LUNG.BREAST_UP.V1_DN   TRUE
#> 18:         c6                                          MEL18_DN.V1_DN   TRUE
#> 19:         c6                                            ATM_DN.V1_UP   TRUE
#> 20:         c6                                      KRAS.LUNG_UP.V1_DN   TRUE
#> 21:         c2                     YAMASHITA_LIVER_CANCER_STEM_CELL_DN   TRUE
#> 22:         c2           CHARAFE_BREAST_CANCER_BASAL_VS_MESENCHYMAL_DN   TRUE
#> 23:         c2                                  BIOCARTA_AGPCR_PATHWAY   TRUE
#> 24:         c2          GUTIERREZ_WALDENSTROEMS_MACROGLOBULINEMIA_1_UP   TRUE
#> 25:         c2                   BOYAULT_LIVER_CANCER_SUBCLASS_G123_DN   TRUE
#> 26:         c6                                        KRAS.50_UP.V1_UP   TRUE
#> 27:         c2                         BURTON_ADIPOGENESIS_PEAK_AT_2HR   TRUE
#> 28:         c2                        HUPER_BREAST_BASAL_VS_LUMINAL_UP   TRUE
#> 29:         c2 TURASHVILI_BREAST_LOBULAR_CARCINOMA_VS_DUCTAL_NORMAL_DN   TRUE
#>     collection                                                    name active
#>          <ord>                                                  <char> <lgcl>
#>         N     n         pval         padj padj.by.collection n.sig n.neutral
#>     <int> <int>        <num>        <num>              <num> <int>     <int>
#>  1:    14    14 9.408031e-09 2.383368e-07       9.408031e-08     8         6
#>  2:   151   149 7.344988e-42 5.582191e-40       2.203497e-40    80        69
#>  3:    56    55 1.103231e-10 4.192278e-09       1.654846e-09    18        37
#>  4:    45    42 5.730276e-06 8.710020e-05       3.438166e-05     2        40
#>  5:   200   175 2.628455e-04 2.853751e-03       5.256910e-03    30       145
#>  6:   196   155 3.581083e-02 9.384907e-02       9.310816e-02    21       134
#>  7:   200   191 3.308409e-03 1.795994e-02       2.205606e-02    11       180
#>  8:    23    23 5.191257e-02 1.195562e-01       9.733607e-02     2        21
#>  9:   172   170 1.279319e-03 9.722825e-03       4.797447e-03    10       160
#> 10:    50    50 3.006263e-02 9.384907e-02       9.310816e-02     7        43
#> 11:   240   233 3.243087e-02 9.384907e-02       6.486174e-02    28       205
#> 12:   200   183 3.287043e-02 9.384907e-02       9.310816e-02    22       161
#> 13:   289   189 2.093147e-02 7.301701e-02       9.159151e-02    23       166
#> 14:   200   176 1.171491e-03 9.722825e-03       1.171491e-02    17       159
#> 15:   145    90 1.915511e-02 7.301701e-02       9.159151e-02     9        81
#> 16:   428   356 1.050618e-03 9.722825e-03       4.502646e-03    63       293
#> 17:   145    96 1.987980e-02 7.301701e-02       9.159151e-02    11        85
#> 18:   148   125 3.490748e-02 9.384907e-02       9.310816e-02    17       108
#> 19:   146    89 1.671038e-02 7.301701e-02       9.159151e-02    13        76
#> 20:   145    98 2.113650e-02 7.301701e-02       9.159151e-02    13        85
#> 21:    76    47 1.540847e-02 7.301701e-02       3.555800e-02     5        42
#> 22:    50    45 1.637298e-03 1.089185e-02       5.159296e-03     7        38
#> 23:    13    11 2.694636e-02 8.904015e-02       5.774220e-02     1        10
#> 24:     9     9 8.344120e-03 4.227687e-02       2.086030e-02     2         7
#> 25:    51    41 2.693973e-05 3.412366e-04       1.346986e-04     9        32
#> 26:    48    33 1.778283e-02 7.301701e-02       9.159151e-02    10        23
#> 27:    51    50 1.719765e-03 1.089185e-02       5.159296e-03     9        41
#> 28:    54    52 2.111603e-03 1.234475e-02       5.758916e-03     9        43
#> 29:    91    91 9.757623e-07 1.853948e-05       7.318218e-06    20        71
#>         N     n         pval         padj padj.by.collection n.sig n.neutral
#>     <int> <int>        <num>        <num>              <num> <int>     <int>
#>      n.up n.down n.sig.up n.sig.down mean.logFC mean.logFC.trim     mean.t
#>     <int>  <int>    <int>      <int>      <num>           <num>      <num>
#>  1:    14      0        8          0  1.9053048       1.9201957  2.6611296
#>  2:   147      2       80          0  1.8402359       1.8310993  2.8107308
#>  3:    47      8       18          0  1.0281693       0.9533016  1.7020744
#>  4:    35      7        2          0  0.3510180       0.3433814  1.2752435
#>  5:   116     59       20         10  0.3908602       0.3337944  0.6218071
#>  6:    98     57        9         12  0.2605313       0.2896170  0.3003986
#>  7:   130     61        4          7  0.2774772       0.2685242  0.4590224
#>  8:    14      9        2          0  0.3419176       0.2139964  0.6136099
#>  9:   104     66        6          4  0.1972357       0.1685089  0.5338254
#> 10:    31     19        5          2  0.2699208       0.1658724  0.4860196
#> 11:   135     98       20          8  0.2188470       0.1531049  0.2752792
#> 12:    75    108        2         20 -0.2631505      -0.2306874 -0.6154836
#> 13:    77    112        1         22 -0.3056071      -0.2627638 -0.6498984
#> 14:    57    119        0         17 -0.3241097      -0.2890946 -0.8570488
#> 15:    31     59        0          9 -0.3264680      -0.3098887 -0.7466725
#> 16:   137    219        5         58 -0.3405691      -0.3268273 -0.7889832
#> 17:    36     60        1         10 -0.3470210      -0.3396559 -0.7333986
#> 18:    48     77        1         16 -0.3763352      -0.3524855 -0.6472540
#> 19:    29     60        1         12 -0.4379007      -0.4117162 -0.7610840
#> 20:    34     64        1         12 -0.4812555      -0.4309138 -0.7247246
#> 21:    16     31        0          5 -0.5284611      -0.4369603 -0.8993549
#> 22:    13     32        0          7 -0.4568979      -0.4711493 -1.1350462
#> 23:     3      8        0          1 -0.4953724      -0.5044566 -1.3726441
#> 24:     2      7        0          2 -0.6026479      -0.6026479 -1.7435775
#> 25:    11     30        0          9 -0.7685127      -0.6500474 -1.5021451
#> 26:     9     24        2          8 -0.5773156      -0.6725080 -0.9812401
#> 27:    14     36        1          8 -0.7654912      -0.7382118 -1.0964370
#> 28:    15     37        1          8 -0.6465576      -0.7566993 -1.0664421
#> 29:    24     67        0         20 -0.9569427      -0.9387142 -1.3803432
#>      n.up n.down n.sig.up n.sig.down mean.logFC mean.logFC.trim     mean.t
#>     <int>  <int>    <int>      <int>      <num>           <num>      <num>
#>     mean.t.trim Direction  rank
#>           <num>    <char> <int>
#>  1:   2.7229413        Up     3
#>  2:   2.8715625        Up     1
#>  3:   1.8177360        Up     2
#>  4:   1.2959816        Up     5
#>  5:   0.7146269        Up     7
#>  6:   0.3953663        Up    29
#>  7:   0.5821929        Up    14
#>  8:   0.4334291        Up    33
#>  9:   0.5781200        Up    10
#> 10:   0.5228263        Up    24
#> 11:   0.2928013        Up    25
#> 12:  -0.5800135      Down    26
#> 13:  -0.5144962      Down    21
#> 14:  -0.7581619      Down     9
#> 15:  -0.5891867      Down    19
#> 16:  -0.7050304      Down     8
#> 17:  -0.6149464      Down    20
#> 18:  -0.5527480      Down    27
#> 19:  -0.7435324      Down    17
#> 20:  -0.6159783      Down    22
#> 21:  -0.8207050      Down    16
#> 22:  -1.0681036      Down    11
#> 23:  -1.3765614      Down    23
#> 24:  -1.7435775      Down    15
#> 25:  -1.5212565      Down     6
#> 26:  -1.1545970      Down    18
#> 27:  -1.1564061      Down    12
#> 28:  -1.0481488      Down    13
#> 29:  -1.2397314      Down     4
#>     mean.t.trim Direction  rank
#>           <num>    <char> <int>