Builds the table of GSEA statistics to present to the user.

Creates a table gene set statistics genereated from a GSEA `method` for all gene sets with an FDR (by collection) less than given threshold.

constructGseaResultTable(
  mg,
  method = sparrow::resultNames(mg)[1L],
  fdr = 0.1,
  prioritize = NULL
)

Arguments

mg

`SparrowResult` object

method

the method to show statistics for

fdr

the FDR cut off to present statistics for

prioritize

the preferred collections to put at the top of the list. The collection column of the table is turned into a factor for more useful display with datatable's filter. Specifcying collections here will put those collections at the front of the factor levels and therofre prioritize their display in the select dropdown for the filter

Value

a data.table of the statistics that match the filtering criteria. A 0-row data.table is returned if nothing passes.

Examples

sres <- sparrow::exampleSparrowResult()
constructGseaResultTable(sres)
#> Error in `[.data.table`(out, , `:=`(padj.by.collection, p.adjust(pval,     "BH")), by = "collection"): attempt access index 18/18 in VECTOR_ELT