Creates a DT::datatable of geneset level GSEA results for use in shiny bits.

This function is closely tied to the exact output of the GSEA methods from the [sparrow::seas()].

renderGseaResultTableDataTable(x, method, mg, digits = 3)

Arguments

x

The set of GSEA statistics generated from from [constructGseaResultTable()]

method

the GSEA method being used fo rdisplay

mg

The `SparrowResult` object. This is used swap in the URL links for genesets using [sparrow::geneSetURL()].

digits

the number of digits to round numer columns to.

Value

A [DT::datatable()] for display

Examples

sres <- sparrow::exampleSparrowResult()
method <- sparrow::resultNames(sres)[1L]
stable <- constructGseaResultTable(sres, method)
renderGseaResultTableDataTable(stable, method, sres, digits = 2)